+++ install/include/nbody/Simulation.h
@@ -17,9 +17,9 @@ namespace nbody {
   public:
     Simulation() : _system{nullptr}, _name{ generateName() } {}
     Simulation( std::istream &input ) : _system{new System(input)}, _name{ generateName() } {}
-    inline Vector3f getPosition( size_t id ) const{ return _system->getPosition(id); }
+    inline Vector3d getPosition( size_t id ) const{ return _system->getPosition(id); }
     inline size_t numBodies() const{ return _system->numBodies(); }
-    void evolveSystem( int nSteps, float dt );
+    void evolveSystem( int nSteps, double dt );
     void loadRun( std::istream &input );
     void saveRun() const;
   };
