+++ install/include/nbody/System.h
@@ -12,8 +12,8 @@ namespace nbody {
   class System {
     size_t _nBodies;
     Body *_body;
-    float _softFactor = 1e-9f;
-    float _dampingFactor = 1.0f;
+    double _softFactor = 1e-9f;
+    double _dampingFactor = 1.0f;
     System() = delete;
     System( const System &sys ) = delete;
     System& operator=( const System &sys ) = delete;
@@ -24,18 +24,18 @@ namespace nbody {
     System( std::string filename ) : _nBodies{}, _body{nullptr} { readState( filename ); }
     ~System() { delete [] _body; }
     inline size_t numBodies() const{ return _nBodies; }
-    void interactBodies( size_t i, size_t j, float softFactor, Vector3f &acc ) const;
+    void interactBodies( size_t i, size_t j, double softFactor, Vector3d &acc ) const;
     void computeGravitation();
-    void integrateSystem( float dt );
-    Vector3f getPosition( size_t id ) const;
+    void integrateSystem( double dt );
+    Vector3d getPosition( size_t id ) const;
     void readState( std::istream &input );
     void readState( std::string filename );
     void writeState( std::ostream &output ) const;
     void writeState( std::string filename ) const;
     void initRandomState();
-    void update( float dt );
-    void setSoftening( float soft ) { _softFactor = soft; }
-    void setDamping( float damp ) { _dampingFactor = damp; }
+    void update( double dt );
+    void setSoftening( double soft ) { _softFactor = soft; }
+    void setDamping( double damp ) { _dampingFactor = damp; }
   };
 
 } // namespace nbody
